+++ CHANGELOG.rst
@@ -70,6 +70,10 @@ time or change existing behaviors in order to make them less surprising/more use
   namespace in which doctests run.
   Thanks `@milliams`_ for the complete PR (`#1428`_).
 
+* New ``--doctest-report`` option available to change the output format of diffs
+  when running (failing) doctests (implements `#1749`_).
+  Thanks `@hartym`_ for the PR.
+
 * New ``name`` argument to ``pytest.fixture`` decorator which allows a custom name
   for a fixture (to solve the funcarg-shadowing-fixture problem).
   Thanks `@novas0x2a`_ for the complete PR (`#1444`_).
@@ -317,6 +321,7 @@ time or change existing behaviors in order to make them less surprising/more use
 .. _#1684: https://github.com/pytest-dev/pytest/pull/1684
 .. _#1723: https://github.com/pytest-dev/pytest/pull/1723
 .. _#1539: https://github.com/pytest-dev/pytest/issues/1539
+.. _#1749: https://github.com/pytest-dev/pytest/issues/1749
 
 .. _@DRMacIver: https://github.com/DRMacIver
 .. _@RedBeardCode: https://github.com/RedBeardCode
@@ -331,6 +336,7 @@ time or change existing behaviors in order to make them less surprising/more use
 .. _@fengxx: https://github.com/fengxx
 .. _@flub: https://github.com/flub
 .. _@graingert: https://github.com/graingert
+.. _@hartym: https://github.com/hartym
 .. _@kalekundert: https://github.com/kalekundert
 .. _@kvas-it: https://github.com/kvas-it
 .. _@marscher: https://github.com/marscher